フォルダ移動した

+++ Classes/GameManager.cpp
@@ -54,7 +54,7 @@ int GameManager::SaveScore = 0;//スコアを保存する
 /////////////////場所////////////////////////
 int GameManager::DistanceMaxDigit = 0;//最大桁数
 int GameManager::Digit = 0;//桁数
-int GameManager::Distance = 0;//距離
+float GameManager::Distance = 0;//距離
 
 int GameManager::SaveDistance = 0;//距離を保存する
 int GameManager::SpriteCnt = 0;//何個目のスプライトかカウントする
