+++ Classes/GameManager.h
@@ -73,7 +73,7 @@ public:
 	const static int DISTANCE_MAX_NUMBER;//スコアのスプライトの数
 	static int DistanceMaxDigit;//最大桁数
 	static int Digit;//桁数
-	static int Distance;//距離
+	static float Distance;//距離
 	static int SaveDistance;//距離を保存する
 	static int SpriteCnt;//何個目のスプライトかカウントする
 	static bool UsedFlag;//この関数がすでに呼ばれたか
